Move bevy's stuff into another module
