+++ src/player.rs
@@ -1,4 +1,4 @@
-use crate::collisions::colliders::{ColliderBundle, ColliderComponent};
+use crate::collisions::plugin::{ColliderBundle, ColliderComponent};
 use crate::collisions::shapes::ColliderShapeType;
 use crate::collisions::store::{ColliderIdResolver, ColliderStore};
 use crate::enemy::Enemy;
